Fix yml squashed (#22)
