+++ assembly/pom.xml
@@ -5,7 +5,7 @@
   <name>Log4JNA Assembly</name>
   <packaging>pom</packaging>
   <parent>
-    <groupId>org.dblock</groupId>
+    <groupId>org.dblock.log4jna</groupId>
     <version>2.0-SNAPSHOT</version>
     <artifactId>log4jna</artifactId>
     <relativePath>../</relativePath>
