+++ assembly/src/main/resources/log4jna-assembly.xml
@@ -32,19 +32,28 @@
       </includes>
     </fileSet>
     <fileSet>
-      <directory>..\demo\target\site</directory>
+      <directory>..\assembly\target\site</directory>
       <filtered>false</filtered>
       <useDefaultExcludes>true</useDefaultExcludes>
-      <outputDirectory>doc\demo</outputDirectory>
+      <outputDirectory>doc\maven-site\assembly</outputDirectory>
       <includes>
         <include>**\*.*</include>
       </includes>
     </fileSet>
     <fileSet>
-      <directory>..\log4jna\target\site</directory>
+      <directory>..\demo\target\site</directory>
       <filtered>false</filtered>
       <useDefaultExcludes>true</useDefaultExcludes>
-      <outputDirectory>doc\log4jna-api</outputDirectory>
+      <outputDirectory>doc\maven-site\demo</outputDirectory>
+      <includes>
+        <include>**\*.*</include>
+      </includes>
+    </fileSet>
+    <fileSet>
+      <directory>..\log4jna-api\target\site</directory>
+      <filtered>false</filtered>
+      <useDefaultExcludes>true</useDefaultExcludes>
+      <outputDirectory>doc\maven-site\log4jna-api</outputDirectory>
       <includes>
         <include>**/*.*</include>
       </includes>
@@ -53,7 +62,7 @@
       <directory>..\target\site</directory>
       <filtered>false</filtered>
       <useDefaultExcludes>true</useDefaultExcludes>
-      <outputDirectory>doc\</outputDirectory>
+      <outputDirectory>doc\maven-site</outputDirectory>
       <includes>
         <include>**\*.*</include>
       </includes>
