+++ assembly/src/main/resources/sources-assembly.xml
