+++ demo/pom.xml
@@ -2,7 +2,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>org.dblock</groupId>
+    <groupId>org.dblock.log4jna</groupId>
     <version>2.0-SNAPSHOT</version>
     <artifactId>log4jna</artifactId>
     <relativePath>../</relativePath>
@@ -60,7 +60,7 @@
   </build>
   <dependencies>
     <dependency>
-      <groupId>org.dblock</groupId>
+      <groupId>org.dblock.log4jna</groupId>
       <artifactId>log4jna-api</artifactId>
       <version>2.0-SNAPSHOT</version>
     </dependency>
