+++ pom.xml
@@ -1,7 +1,7 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.dblock</groupId>
+  <groupId>org.dblock.log4jna</groupId>
   <artifactId>log4jna</artifactId>
   <version>2.0-SNAPSHOT</version>
   <packaging>pom</packaging>
@@ -143,6 +143,17 @@
           </execution>
         </executions>
       </plugin>
+      <plugin>
+        <groupId>org.sonatype.plugins</groupId>
+        <artifactId>nexus-staging-maven-plugin</artifactId>
+        <version>1.6.7</version>
+        <extensions>true</extensions>
+        <configuration>
+          <serverId>osshr</serverId>
+          <nexusUrl>https://oss.sonatype.org/</nexusUrl>
+          <autoReleaseAfterClose>false</autoReleaseAfterClose>
+        </configuration>
+      </plugin>
     </plugins>
   </build>
   <properties>
@@ -154,6 +165,8 @@
     <mavencompilerplugin.version>3.5.1</mavencompilerplugin.version>
     <mavenjavadocplugin.version>2.10.4</mavenjavadocplugin.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <project.build.resourceEncoding>UTF-8</project.build.resourceEncoding>
+    <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
   </properties>
   <dependencyManagement>
     <dependencies>
@@ -183,23 +196,22 @@
     </dependency>
   </dependencies>
   <distributionManagement>
+    <!-- We have to find a location for the site -->
+    <!-- 
     <site>
       <id>sites</id>
       <name>CGTCA Project Sites</name>
       <url>scp://httpd.cgtca.ca/var/www/sites/log4jna/</url>
     </site>
-    <repository>
-      <id>CGTCANexus</id>
-      <layout>default</layout>
-      <name>CGTCA Repo - Nexus Server Releases</name>
-      <url>http://repo.cgtca.ca:8081/nexus/content/repositories/releases/</url>
-    </repository>
+    -->
     <snapshotRepository>
-      <id>CGTCANexus</id>
-      <layout>default</layout>
-      <name>CGTCA Repo - Nexus Server Snapshots</name>
-      <url>http://repo.cgtca.ca:8081/nexus/content/repositories/snapshots/</url>
+      <id>osshr</id>
+      <url>https://oss.sonatype.org/content/repositories/snapshots</url>
     </snapshotRepository>
+    <repository>
+      <id>osshr</id>
+      <url>https://oss.sonatype.org/service/local/staging/deploy/maven2</url>
+    </repository>
   </distributionManagement>
   <modules>
     <module>win32dll</module>
