+++ win32dll/pom.xml
@@ -5,7 +5,7 @@
   <packaging>pom</packaging>
   <name>Log4JNA Win32dll</name>
   <parent>
-    <groupId>org.dblock</groupId>
+    <groupId>org.dblock.log4jna</groupId>
     <artifactId>log4jna</artifactId>
     <version>2.0-SNAPSHOT</version>
     <relativePath>../</relativePath>
