+++ win32dll/src/EventLogCategories.h
