+++ win32dll/src/EventLogCategories.mc
