+++ win32dll/src/EventLogCategories.rc
