+++ win32dll/src/Win32EventLogAppender.rc
