+++ win32dll/src/version.rc
